requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0
pynput>=1.7.0
pyyaml>=5.4.0
RPi.GPIO>=0.7.0; platform_system == "Linux"
